Added Utility files from main.js
Seperated utility files from main.js.
