+++ edgecompute/examples/authentication/ja3-fingerprinting/bundle.json
@@ -1,4 +1,4 @@
 {
-	"edgeworker-version":"1.0",
-	"description":"New version 1.0 - Calculates JA3 Fingerprint using EdgeWorkers."
+    "edgeworker-version": "1.0",
+    "description": "New version 1.0 - Calculates JA3 Fingerprint using EdgeWorkers."
 }
